Adicionando validacao no endpoint de deletar tweet

+++ twitter-api/src/tweets/tweets.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+	Injectable,
+	NotFoundException,
+	UnauthorizedException,
+} from '@nestjs/common';
 import { TweetsRepository } from './tweets.repository';
 import { CriarTweetDto } from './dto/criar-tweet.dto';
 import { Tweet } from './tweet.entity';
@@ -25,9 +29,10 @@ export class TweetsService {
 		return tweet;
 	}
 
-	async deletarTweet(idTweet: string, idUsuario: string) {
-		//todo: comparar ids
-		idUsuario;
+	async deletarTweet(idTweet: string, usuarioId: string) {
+		const tweet = await this.encontrarTweetPeloId(idTweet);
+
+		if (tweet.usuarioId !== usuarioId) throw new UnauthorizedException();
 		const resultado = await this.tweetsRepository.delete({ id: idTweet });
 		if (resultado.affected === 0) throw new NotFoundException();
 	}
